[ref-perf] move data-filter options channel reading up to fix bugs related to initialization of the hook
